moviedb-106 Move `database.FindMovieDict` to `config.py`.
